requirements.txt
python-decouple
django
djangorestframework
drf-nested-routers
mysqlclient
redis
django_redis
drf-yasg
bcrypt
celery